+++ MainSystem/Bank/BankTransfers_Update.cs
@@ -101,47 +101,47 @@ namespace MainSystem
                     t200.Text = dr["a200"].ToString();
                     if (dr["a200"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[0] = Convert.ToInt16(dr["a200"].ToString());
+                        arrPaidMoneyPlus[0] = Convert.ToInt32(dr["a200"].ToString());
                     }
                     t100.Text = dr["a100"].ToString();
                     if (dr["a100"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[1] = Convert.ToInt16(dr["a100"].ToString());
+                        arrPaidMoneyPlus[1] = Convert.ToInt32(dr["a100"].ToString());
                     }
                     t50.Text = dr["a50"].ToString();
                     if (dr["a50"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[2] = Convert.ToInt16(dr["a50"].ToString());
+                        arrPaidMoneyPlus[2] = Convert.ToInt32(dr["a50"].ToString());
                     }
                     t20.Text = dr["a20"].ToString();
                     if (dr["a20"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[3] = Convert.ToInt16(dr["a20"].ToString());
+                        arrPaidMoneyPlus[3] = Convert.ToInt32(dr["a20"].ToString());
                     }
                     t10.Text = dr["a10"].ToString();
                     if (dr["a10"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[4] = Convert.ToInt16(dr["a10"].ToString());
+                        arrPaidMoneyPlus[4] = Convert.ToInt32(dr["a10"].ToString());
                     }
                     t5.Text = dr["a5"].ToString();
                     if (dr["a5"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[5] = Convert.ToInt16(dr["a5"].ToString());
+                        arrPaidMoneyPlus[5] = Convert.ToInt32(dr["a5"].ToString());
                     }
                     t1.Text = dr["a1"].ToString();
                     if (dr["a1"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[6] = Convert.ToInt16(dr["a1"].ToString());
+                        arrPaidMoneyPlus[6] = Convert.ToInt32(dr["a1"].ToString());
                     }
                     tH.Text = dr["aH"].ToString();
                     if (dr["aH"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[7] = Convert.ToInt16(dr["aH"].ToString());
+                        arrPaidMoneyPlus[7] = Convert.ToInt32(dr["aH"].ToString());
                     }
                     tQ.Text = dr["aQ"].ToString();
                     if (dr["aQ"].ToString() != "")
                     {
-                        arrPaidMoneyPlus[8] = Convert.ToInt16(dr["aQ"].ToString());
+                        arrPaidMoneyPlus[8] = Convert.ToInt32(dr["aQ"].ToString());
                     }
                 }
                 dr.Close();
@@ -391,15 +391,15 @@ namespace MainSystem
                             MySqlDataReader dr = com2.ExecuteReader();
                             while (dr.Read())
                             {
-                                arrOFPhaatPlus[0] = Convert.ToInt16(dr["a200"]);
-                                arrOFPhaatPlus[1] = Convert.ToInt16(dr["a100"]);
-                                arrOFPhaatPlus[2] = Convert.ToInt16(dr["a50"]);
-                                arrOFPhaatPlus[3] = Convert.ToInt16(dr["a20"]);
-                                arrOFPhaatPlus[4] = Convert.ToInt16(dr["a10"]);
-                                arrOFPhaatPlus[5] = Convert.ToInt16(dr["a5"]);
-                                arrOFPhaatPlus[6] = Convert.ToInt16(dr["a1"]);
-                                arrOFPhaatPlus[7] = Convert.ToInt16(dr["aH"]);
-                                arrOFPhaatPlus[8] = Convert.ToInt16(dr["aQ"]);
+                                arrOFPhaatPlus[0] = Convert.ToInt32(dr["a200"]);
+                                arrOFPhaatPlus[1] = Convert.ToInt32(dr["a100"]);
+                                arrOFPhaatPlus[2] = Convert.ToInt32(dr["a50"]);
+                                arrOFPhaatPlus[3] = Convert.ToInt32(dr["a20"]);
+                                arrOFPhaatPlus[4] = Convert.ToInt32(dr["a10"]);
+                                arrOFPhaatPlus[5] = Convert.ToInt32(dr["a5"]);
+                                arrOFPhaatPlus[6] = Convert.ToInt32(dr["a1"]);
+                                arrOFPhaatPlus[7] = Convert.ToInt32(dr["aH"]);
+                                arrOFPhaatPlus[8] = Convert.ToInt32(dr["aQ"]);
                             }
                             dr.Close();
 
@@ -408,15 +408,15 @@ namespace MainSystem
                             MySqlDataReader dr2 = com2.ExecuteReader();
                             while (dr2.Read())
                             {
-                                arrOFPhaatMinus[0] = Convert.ToInt16(dr2["a200"]);
-                                arrOFPhaatMinus[1] = Convert.ToInt16(dr2["a100"]);
-                                arrOFPhaatMinus[2] = Convert.ToInt16(dr2["a50"]);
-                                arrOFPhaatMinus[3] = Convert.ToInt16(dr2["a20"]);
-                                arrOFPhaatMinus[4] = Convert.ToInt16(dr2["a10"]);
-                                arrOFPhaatMinus[5] = Convert.ToInt16(dr2["a5"]);
-                                arrOFPhaatMinus[6] = Convert.ToInt16(dr2["a1"]);
-                                arrOFPhaatMinus[7] = Convert.ToInt16(dr2["aH"]);
-                                arrOFPhaatMinus[8] = Convert.ToInt16(dr2["aQ"]);
+                                arrOFPhaatMinus[0] = Convert.ToInt32(dr2["a200"]);
+                                arrOFPhaatMinus[1] = Convert.ToInt32(dr2["a100"]);
+                                arrOFPhaatMinus[2] = Convert.ToInt32(dr2["a50"]);
+                                arrOFPhaatMinus[3] = Convert.ToInt32(dr2["a20"]);
+                                arrOFPhaatMinus[4] = Convert.ToInt32(dr2["a10"]);
+                                arrOFPhaatMinus[5] = Convert.ToInt32(dr2["a5"]);
+                                arrOFPhaatMinus[6] = Convert.ToInt32(dr2["a1"]);
+                                arrOFPhaatMinus[7] = Convert.ToInt32(dr2["aH"]);
+                                arrOFPhaatMinus[8] = Convert.ToInt32(dr2["aQ"]);
                             }
                             dr2.Close();
                             flag = true;
