+++ MainSystem/Bank/عميل/سحب كاش/BankPullCash_Record.cs
@@ -141,7 +141,7 @@ namespace MainSystem
                             while (dr.Read())
                             {
                                 flag2 = true;
-                                ID = Convert.ToInt16(dr["CustomerReturnBill_ID"].ToString());
+                                ID = Convert.ToInt32(dr["CustomerReturnBill_ID"].ToString());
                                 TypeBuy = dr["Type_Buy"].ToString();
                                 billDate = Convert.ToDateTime(dr["Date"].ToString());
                                 returnInfo = dr["ReturnInfo"].ToString();
@@ -187,12 +187,12 @@ namespace MainSystem
 
                                 if (dr["Customer_ID"].ToString() != "")
                                 {
-                                    customerID = Convert.ToInt16(dr["Customer_ID"].ToString());
+                                    customerID = Convert.ToInt32(dr["Customer_ID"].ToString());
                                 }
 
                                 if (dr["Client_ID"].ToString() != "")
                                 {
-                                    clientID = Convert.ToInt16(dr["Client_ID"].ToString());
+                                    clientID = Convert.ToInt32(dr["Client_ID"].ToString());
                                 }
                             }
                             dr.Close();
@@ -646,7 +646,7 @@ namespace MainSystem
                                 com.Parameters.Add("@r1", MySqlDbType.Int16, 11).Value = arrRestMoney[6];
                                 com.Parameters.Add("@rH", MySqlDbType.Int16, 11).Value = arrRestMoney[7];
                                 com.Parameters.Add("@rQ", MySqlDbType.Int16, 11).Value = arrRestMoney[8];
-                                com.Parameters.Add("@Transition_ID", MySqlDbType.Int16, 11).Value = Convert.ToInt16(TransitionID);
+                                com.Parameters.Add("@Transition_ID", MySqlDbType.Int16, 11).Value = Convert.ToInt32(TransitionID);
                                 com.ExecuteNonQuery();
                                 flagCategoriesSuccess = false;
                                 //////////////////////
@@ -755,15 +755,15 @@ namespace MainSystem
                             MySqlDataReader dr = com2.ExecuteReader();
                             while (dr.Read())
                             {
-                                arrOFPhaat[0] = Convert.ToInt16(dr["a200"]);
-                                arrOFPhaat[1] = Convert.ToInt16(dr["a100"]);
-                                arrOFPhaat[2] = Convert.ToInt16(dr["a50"]);
-                                arrOFPhaat[3] = Convert.ToInt16(dr["a20"]);
-                                arrOFPhaat[4] = Convert.ToInt16(dr["a10"]);
-                                arrOFPhaat[5] = Convert.ToInt16(dr["a5"]);
-                                arrOFPhaat[6] = Convert.ToInt16(dr["a1"]);
-                                arrOFPhaat[7] = Convert.ToInt16(dr["aH"]);
-                                arrOFPhaat[8] = Convert.ToInt16(dr["aQ"]);
+                                arrOFPhaat[0] = Convert.ToInt32(dr["a200"]);
+                                arrOFPhaat[1] = Convert.ToInt32(dr["a100"]);
+                                arrOFPhaat[2] = Convert.ToInt32(dr["a50"]);
+                                arrOFPhaat[3] = Convert.ToInt32(dr["a20"]);
+                                arrOFPhaat[4] = Convert.ToInt32(dr["a10"]);
+                                arrOFPhaat[5] = Convert.ToInt32(dr["a5"]);
+                                arrOFPhaat[6] = Convert.ToInt32(dr["a1"]);
+                                arrOFPhaat[7] = Convert.ToInt32(dr["aH"]);
+                                arrOFPhaat[8] = Convert.ToInt32(dr["aQ"]);
                             }
                             flag = true;
                         }
@@ -1436,7 +1436,7 @@ namespace MainSystem
                         productQ = Convert.ToDouble(dr["TotalMeter"]);
 
                         storageQ += productQ;
-                        id = Convert.ToInt16(dr2["Storage_ID"]);
+                        id = Convert.ToInt32(dr2["Storage_ID"]);
                         q = "update storage set Total_Meters=" + storageQ + " where Storage_ID=" + id;
                         MySqlCommand comm = new MySqlCommand(q, dbconnection);
                         comm.ExecuteNonQuery();
@@ -1459,7 +1459,7 @@ namespace MainSystem
                         productQ = Convert.ToDouble(dr["TotalMeter"]);
 
                         storageQ += productQ;
-                        id = Convert.ToInt16(dr2["Storage_ID"]);
+                        id = Convert.ToInt32(dr2["Storage_ID"]);
                         q = "update storage set Total_Meters=" + storageQ + " where Storage_ID=" + id;
                         MySqlCommand comm = new MySqlCommand(q, dbconnection);
                         comm.ExecuteNonQuery();
@@ -1482,7 +1482,7 @@ namespace MainSystem
                     productQ = Convert.ToDouble(dr["TotalMeter"]);
 
                     storageQ += productQ;
-                    id = Convert.ToInt16(dr3["StorageTaxesID"]);
+                    id = Convert.ToInt32(dr3["StorageTaxesID"]);
                     q = "update storage_taxes set Total_Meters=" + storageQ + " where StorageTaxesID=" + id;
                     MySqlCommand comm = new MySqlCommand(q, dbconnection);
                     comm.ExecuteNonQuery();
